refactor: removed unnecessary foler nesting in auth module
