+++ src/app/core/custom-http.interceptor.ts
@@ -60,5 +60,3 @@ export class CustomHttpInterceptor implements HttpInterceptor {
     return req.clone({ url });
   }
 }
-
-////
