+++ src/app/features/auth/index.ts
@@ -1,6 +1,6 @@
 export * from './auth.module';
-export * from './auth/login/auth.component';
-export * from './auth/auth.service';
-export * from './auth/auth-validators';
-export * from './auth/auth.types';
-export * from './auth/handle-api-error.service';
+export * from './login/auth.component';
+export * from './auth.service';
+export * from './auth-validators';
+export * from './auth.types';
+export * from './handle-api-error.service';
